+++ nullaway/src/main/java/com/uber/nullaway/NullabilityUtil.java
@@ -44,6 +44,7 @@ import com.sun.tools.javac.util.JCDiagnostic;
 import java.util.stream.Stream;
 import javax.annotation.Nullable;
 import javax.lang.model.element.AnnotationMirror;
+import org.checkerframework.javacutil.AnnotationUtils;
 
 /** Helpful utility methods for nullability analysis. */
 public class NullabilityUtil {
@@ -158,6 +159,23 @@ public class NullabilityUtil {
     // for methods, we care about annotations on the return type, not on the method type itself
     Stream<? extends AnnotationMirror> typeUseAnnotations = getTypeUseAnnotations(symbol);
     return Stream.concat(symbol.getAnnotationMirrors().stream(), typeUseAnnotations);
+  }
+
+  /**
+   * Retrieve the string value inside an annotation without statically depending on the type.
+   *
+   * @param annotName Annotation name to retrieve it's value.
+   * @param methodSymbol A method which has an @Contract annotation.
+   * @return The string value spec inside the annotation.
+   */
+  public static @Nullable String getAnnotationValue(
+      Symbol.MethodSymbol methodSymbol, String annotName) {
+    AnnotationMirror annot =
+        AnnotationUtils.getAnnotationByName(methodSymbol.getAnnotationMirrors(), annotName);
+    if (annot == null) {
+      return null;
+    }
+    return AnnotationUtils.getElementValue(annot, "value", String.class, true);
   }
 
   /**
